+++ smtk/attribute/Testing/basicAttributeDerivationTest.cxx
@@ -82,7 +82,7 @@ int main()
   smtk::attribute::ValueItemPtr vcomp;
   smtk::attribute::ItemPtr comp;
 
-  int i, n = att->numberOfItems();
+  int i, n = static_cast<int>(att->numberOfItems());
   std::cout << "Items of testAtt:\n";
   for (i = 0; i < n; i++)
     {
